switch to using nomonouns/assets npm package
switch to using the nomonouns/assets npm package as it is currently being updated when traits are added to nouns in order for nomonouns to work. nomo also needs fomo to work, so aligning the package is probably a good idea. nomo now uses our own npm package so we can push the update to npm right when the trait proposals are executed at nouns.